Move permissions from group1 to own group
